+++ CHANGELOG.md
@@ -5,6 +5,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Not released:
 
+* Fix line with tension calculations
+
 ## 4.1.2 - 2020-01-08
 
 * Fix possible `NaN` in content calculations